+++ activejob/lib/active_job/queue_adapter.rb
@@ -37,12 +37,6 @@ module ActiveJob
           else
             if queue_adapter?(name_or_adapter_or_class)
               name_or_adapter_or_class
-            elsif queue_adapter_class?(name_or_adapter_or_class)
-              ActiveSupport::Deprecation.warn "Passing an adapter class is deprecated " \
-                "and will be removed in Rails 5.1. Please pass an adapter name " \
-                "(.queue_adapter = :#{name_or_adapter_or_class.name.demodulize.remove('Adapter').underscore}) " \
-                "or an instance (.queue_adapter = #{name_or_adapter_or_class.name}.new) instead."
-              name_or_adapter_or_class.new
             else
               raise ArgumentError
             end
@@ -53,10 +47,6 @@ module ActiveJob
 
         def queue_adapter?(object)
           QUEUE_ADAPTER_METHODS.all? { |meth| object.respond_to?(meth) }
-        end
-
-        def queue_adapter_class?(object)
-          object.is_a?(Class) && QUEUE_ADAPTER_METHODS.all? { |meth| object.public_method_defined?(meth) }
         end
     end
   end